requirements.txt
supabase
requests
beautifulsoup4
google-generativeai
python-dotenv
